Add files via upload
A University Management System Project in Java that is built with the utmost precision using NetBeans and MySQL Database as the System's Back-End. It was designed specifically for the University and its affiliated institutions to carry out, observe, and assess complex operations such as Storing student information, Storing Teacher information, generating fee challans and many more with unparalleled efficiency.